Extensive updates of idtracker.
ballot sets are now (hopefully) treated properly, except for the
row coloring.  They're also included in the "view_id" view.

Search is re-done, it turns out that using draft__<anything> means
that RFCs won't be matched, so build lists of possibly-matching
RFCs and I-Ds and pass them through.  This applies to filename,
group and rfc number.


git-svn-id: http://svn.tools.ietf.org/svn/tools/ietfdb/trunk@190 7b24d068-2d4e-4fce-9bd7-cbd2762980b0