Rename "initial" to "initial_members" so that it doesn't shadow the form attribute "initial" (as in initial data)
Render the form field-by-field for MailingListForm, so that
we can render the mailing list name conditionally as a text input
box or as a hidden field with the value in the page.  We can also
insert the domain name here too.  (Note that this means that
hidden fields in this form other than domain name have to be
treated specially!)


git-svn-id: http://svn.tools.ietf.org/svn/tools/ietfdb/trunk@270 7b24d068-2d4e-4fce-9bd7-cbd2762980b0