Merged [5570] from rjsparks@nostrum.com:
Moves the interface to edit shepherds and sheperd writeups out of the 
wgtracker interface and into the doc views (via idrfc).
Added a little to the simple authorization code (can_edit_) in idrfc.
Added a way to reset the writeup to the current templates (leaving 
extension points to take templates for non-ietf stream documents).
Fixes bug #967.

Implementation notes:
Currently reused EmailsField -- in the long 
run a better approach will be to build a similar PersonField.

This also fixes tickets #834 and #691 (which were effectively duplicates 
of #967)


git-svn-id: http://svn.tools.ietf.org/svn/tools/ietfdb/trunk@5631 7b24d068-2d4e-4fce-9bd7-cbd2762980b0